refactor(htmlcss::svg): split dom/element.rs, extract text shaping, drop dead filter::apply
Continues the Blink-aligned reorganization with the items previously
flagged as regression risk. Reftest score unchanged at 89.50%
(1679 fixtures, exact baseline match).

dom/element.rs split into three:
- dom/element.rs: keeps ElementKind, get_attr, element_kind (pure DOM)
- paint/visibility.rs: is_painted, has_display_none, is_visible_self,
  is_visible_inherited (paint dispatch concerns)
- style/cascade.rs: cascade_property, get_attr_or_style (single funnel
  for property resolution; future Stylo migration target)

Updates the three callers (svg_root_painter, svg_container_painter,
svg_shape_painter) to import from the new locations.

paint/svg_text_painter.rs (2,468 LOC) split into a directory:
- svg_text_painter/mod.rs: orchestrator, attribute flattening, layout
- svg_text_painter/shaping.rs (~190 LOC): SkShaper integration,
  ShapedGlyph, compute_kerned_advances, shape_text
Further text submodules (textPath, layout algorithm) deferred — their
cross-dependencies risk regression for marginal gain.

Removes resources/filter.rs::apply (dead code; the painter applies
filters inline at svg_container_painter.rs). Eliminates the only
resources/ Canvas violation, removing two allowlist entries from
tests/htmlcss_svg_architecture.rs.

README module map updated.
